requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
pillow
httpx
aptos-sdk